Scaffolding Exames, e correçao de erros, tabelas etcc
