+++ HealthWellbeing.csproj
@@ -23,4 +23,8 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="8.0.7" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Views\Exame\" />
+  </ItemGroup>
+
 </Project>
